+++ farhan/tugas_akhir/index.php
@@ -16,6 +16,7 @@
         <a href="./">Barang</a>
         <a href="./barang_masuk.php">Log Masuk</a>
         <a href="./barang_keluar.php">Log Keluar</a>
+        <a href="./aksi/auth/logout.php">Logout</a>
     </div>
     <a href="./form_barang.php">Tambah Data</a>
     <?php if(isset($_GET['pesan'])){
@@ -30,7 +31,9 @@
                 <th>Deskripsi</th>
                 <th>Satuan</th>
                 <th>Stock</th>
-                <th>Aksi</th>
+                <?php if(isset($_SESSION['role']) && $_SESSION['role'] == 'Admin') {?>
+                    <th>Aksi</th>
+                <?php }?>
             </tr>
         </thead>
         <tbody>
@@ -45,9 +48,11 @@
                     <td><?= $d['deskripsi'] ?></td>
                     <td><?= $d['satuan'] ?></td>
                     <td><?= $d['stock'] ?></td>
-                    <td>
-                        <a href="./edit_barang.php?id=<?=$d['id']?>">Edit</a>|<a href="./aksi/hapus_barang.php?id=<?=$d['id']?>">Hapus</a>
-                    </td>
+                    <?php if(isset($_SESSION['role']) && $_SESSION['role'] == 'Admin') {?>
+                        <td>
+                            <a href="./edit_barang.php?id=<?=$d['id']?>">Edit</a>|<a href="./aksi/hapus_barang.php?id=<?=$d['id']?>">Hapus</a>
+                        </td>
+                    <?php }?>
                 </tr>
             <?php
                     $no++;
